health slider+ player die state

+++ Assets/Scripts/vu/Player/TakeDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
  interface  TakeDamage 
 {
-   public void TakeDamage( float damage );
+   public void TakeDamage( int damage );
 }
